Fixes, UX Tweaks, refactor FastAPI error handling and data types (#58)
* Update .gitignore and refactor FastAPI error handling and data types

- Added .vscode/, .cursor/, and .qodo/ to .gitignore to prevent unnecessary files from being tracked.
- Refactored error handling in the FastAPI application to simplify the validation exception handler, removing unnecessary logging of request body details.
- Updated data types in various models and endpoints to use Decimal instead of float for financial values, ensuring better precision in calculations.
- Adjusted related API responses and request handling to accommodate the new data types, enhancing overall data integrity.

* Refactor FastAPI error handling and improve data types across the application

- Cleaned up the validation exception handler in FastAPI to enhance readability.
- Updated various models and API endpoints to consistently use Decimal for financial values, improving precision in calculations.
- Enhanced the unit retrieval process by eager loading tenant details, optimizing performance.
- Implemented search functionality in the RentTracker component to filter rent data based on tenant and property names.
- Adjusted UI elements for better alignment and clarity in the RentTracker and Leases components.

* dash

* Enhance Dashboard and RentTracker components

- Updated the Dashboard component to return a more accurate percentage change when the previous number is zero, accounting for negative values.
- Refactored the RentTracker component to ensure proper handling of property names and financial values, using Number.parseFloat for consistency and improved readability.
- Added a fallback for property names to display "N/A" when not available, enhancing user experience.

---------

+++ Backend/api/accounting/invoices.py
@@ -1,5 +1,6 @@
 import logging
 from datetime import date, datetime
+from decimal import Decimal
 
 from fastapi import APIRouter, Depends, HTTPException, status
 from pydantic import BaseModel
@@ -30,7 +31,7 @@ router = APIRouter()
 # === API Models for Invoices ===
 class InvoiceBase(BaseModel):
     invoice_number: str
-    amount: float
+    amount: Decimal
     description: str
     issue_date: datetime
     due_date: datetime
@@ -42,7 +43,7 @@ class InvoiceCreate(InvoiceBase):
     pass
 
 class InvoiceUpdate(BaseModel):
-    amount: float | None = None
+    amount: Decimal | None = None
     description: str | None = None
     issue_date: datetime | None = None
     due_date: datetime | None = None
